+++ my/theming/hypr-vogix/mynixos.nix
@@ -4,6 +4,6 @@
 
 {
   config = {
-    my.themes.hypr-vogix.enable = lib.mkDefault false; # TODO: move into vogix
+    my.theming.hypr-vogix.enable = lib.mkDefault false; # TODO: move into vogix
   };
 }