+++ my/theming/stylix/default.nix
@@ -1,0 +1,7 @@
+# Stylix static theming - legacy, disabled by default
+# Options are defined in my/theming/options.nix under theming.stylix
+# Theme loading is handled by mkSystem.nix when my.theming.stylix.type = "stylix"
+# This module is intentionally empty as stylix configuration is injected externally
+_:
+
+{ }